RUMM-2616 Remove URLSessionInterceptor
