+++ Tests/DatadogTests/DatadogObjc/DDDatadogTests.swift
@@ -6,8 +6,8 @@
 
 import XCTest
 import TestUtilities
-import DatadogInternal
 
+@testable import DatadogInternal
 @testable import DatadogLogs
 @testable import Datadog
 @testable import DatadogObjc
@@ -38,17 +38,14 @@ class DDDatadogTests: XCTestCase {
 
         XCTAssertTrue(Datadog.isInitialized)
 
-//        let urlSessionInstrumentation = defaultDatadogCore.v1.feature(URLSessionAutoInstrumentation.self)
         let context = try XCTUnwrap(defaultDatadogCore as? DatadogCore).contextProvider.read()
         XCTAssertEqual(context.applicationName, "app-name")
         XCTAssertEqual(context.env, "tests")
-//        XCTAssertNotNil(urlSessionInstrumentation)
 
-//        urlSessionInstrumentation?.swizzler.unswizzle()
         Datadog.flushAndDeinitialize()
 
         XCTAssertNil(defaultDatadogCore.get(feature: DatadogLogsFeature.self))
-        XCTAssertNil(defaultDatadogCore.v1.feature(URLSessionAutoInstrumentation.self))
+        XCTAssertNil(defaultDatadogCore.get(feature: NetworkInstrumentationFeature.self))
     }
 
     // MARK: - Changing Tracking Consent
