+++ Tests/DatadogTests/DatadogObjc/DDNSURLSessionDelegateTests.swift
@@ -6,6 +6,7 @@
 
 import XCTest
 import TestUtilities
+import DatadogInternal
 @testable import Datadog
 @testable import DatadogObjc
